Move getPostTypeTaxonomyArchive to functions
