+++ functions/wordpress/postTypes/getPostTypeTaxonomyArchive.js
